+++ src/library/components/table/editors/custom.tsx
@@ -13,7 +13,9 @@ export const CustomEditor: React.FC<EditorProps> = function CustomEditor({row, c
     getRowKey,
     onSaveQuickEdit,
   } = useContext(ColumnContext);
-  const Editor = (column.editor as CustomColumnEditor<any>).Editor;
+
+  const editorDef = column.editor as CustomColumnEditor<any>;
+  const Editor = editorDef.Editor;
 
   // yes we should be able to count on this
   // earlier validation assures editing only if either
@@ -46,6 +48,23 @@ export const CustomEditor: React.FC<EditorProps> = function CustomEditor({row, c
     }))
   }
 
+  function setValues(data: any) {
+    if (!data || typeof data !== 'object' || Array.isArray(data)) {
+      throw new Error('Invalid data format for setting data');
+    }
+
+    let keys = Object.keys(data);
+    let updateSpec: any = {};
+    for (let key of keys) {
+      updateSpec[key] = { $set: data[key] };
+    }
+    setCtxData(data => update(data, {
+      editData: {
+        [keyValue]: { $auto: updateSpec }
+      }
+    }));
+  }
+
   function autoSave() {
     if (Object.keys(rowData as object).length)
       onSaveQuickEdit({[keyValue]: rowData as any});
@@ -57,7 +76,10 @@ export const CustomEditor: React.FC<EditorProps> = function CustomEditor({row, c
     row={row}
     column={column}
     setValue={onChange}
+    setValues={setValues}
     autoSave={autoSave}
     editMode={editMode}
+    editData={rowData as any}
+    editorOptions={editorDef.editorOptions}
   />;
 }