+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 ## [Unreleased]
 
+## [1.7.0]
+### Added
+- Pass row edit data to custom cell editors - allows altering other editors based on changes to other cells.
+- `setValues` function to custom editors to allow updating multiple columns at once (ex. side effect changes).
+- Filter editor support for custom single field multi value entries (multi-select dropdowns)
+- Expanded `editorOptions` to custom row editors not just filter editors
+
+### Fixed
+- Scroll to top on page change with fixed headers
+- Some styling changes with headers due to fixed headers change
+
 ## [1.6.9]
 ### Fixed
 - Removed some console.logs